Create det_processes.c
This exercise was a simple project in our Operating Systems Course. This was the description of the requirements: 
Design a program using ordinary pipes in which one process sends a sorted array to a second process, and the second process reverses the array and sends it back to the first process. For example, if the first process sends the message 5 6 7 8 9, the second process will return 9 8 7 6 5. This will require using two pipes, one for sending the original message from the first to the second process and the other for sending the modified message from the second to the first process.